Conflicts:
	code/game/mecha/mecha.dm
	code/modules/mob/living/silicon/ai/ai.dm

+++ code/modules/paperwork/paper.dm
@@ -335,15 +335,17 @@
 	burn(1, 50)
 
 /obj/item/weapon/paper/proc/burn(var/showmsg, var/burntime)
+	if (burning)
+		return
 	if(showmsg)
-		src.visible_message("<span class='warning'>[src] catches on fire!</span>")
+		src.visible_message("<span class='warning'>[src] catches on fire.</span>")
 	burning = 1
 	icon_state = "paper_onfire"
 	info = "[stars(info)]"
-	sleep(burntime) //7 seconds
-	src.visible_message("<span class='danger'>[src] burns away, leaving behind a pile of ashes.</span>")
-	new /obj/effect/decal/cleanable/ash(src.loc)
-	qdel(src)
+	spawn(burntime) //7 seconds
+		src.visible_message("<span class='warning'>[src] burns away, leaving behind a pile of ashes.</span>")
+		new /obj/effect/decal/cleanable/ash(src.loc)
+		qdel(src)
 
 /*
  * Premade paper
